// admin dashboard/kaaval-backend/src/auth/roles.enum.ts
export enum Role {
  SUPER_ADMIN = 'super_admin',
  TRAFFIC_ADMIN = 'traffic_admin',
  DEV_ADMIN = 'dev_admin',
  COLACHEL_ADMIN = 'colachel_admin',
  MARTHANDAM_ADMIN = 'marthandam_admin',
  NAGERCOIL_ADMIN = 'nagercoil_admin',
  KANYAKUMARI_ADMIN = 'kanyakumari_admin',
  THUCKALAY_ADMIN = 'thuckalay_admin',
}